+++ pssh/clients/base/parallel.py
@@ -88,12 +88,7 @@ class BaseParallelSSHClient(object):
     def _check_host_config(self):
         if self.host_config is None:
             return
-        host_len = 0
-        try:
-            host_len = len(self.hosts)
-        except TypeError:
-            # Generator
-            return
+        host_len = len(self.hosts)
         if host_len != len(self.host_config):
             raise ValueError(
                 "Host config entries must match number of hosts if provided. "
@@ -169,8 +164,10 @@ class BaseParallelSSHClient(object):
         finished_shells = [g.get() for g in finished]
         unfinished_shells = list(set(shells).difference(set(finished_shells)))
         if len(unfinished_shells) > 0:
-            raise Timeout("Timeout of %s sec(s) reached with commands "
-                          "still running", timeout, finished_shells, unfinished_shells)
+            raise Timeout(
+                "Timeout of %s sec(s) reached with commands still running",
+                timeout, finished_shells, unfinished_shells,
+            )
 
     def run_command(self, command, user=None, stop_on_errors=True,
                     host_args=None, use_pty=False, shell=None,
@@ -354,8 +351,10 @@ class BaseParallelSSHClient(object):
         if unfinished_cmds:
             finished_output = self.get_last_output(cmds=finished_cmds)
             unfinished_output = list(set.difference(set(output), set(finished_output)))
-            raise Timeout("Timeout of %s sec(s) reached with commands "
-                          "still running", timeout, finished_output, unfinished_output)
+            raise Timeout(
+                "Timeout of %s sec(s) reached with commands still running",
+                timeout, finished_output, unfinished_output,
+            )
 
     def _join(self, host_out, consume_output=False, timeout=None,
               encoding="utf-8"):