+++ pssh/clients/base/single.py
@@ -323,7 +323,10 @@ class BaseSSHClient(object):
     def _password_auth(self):
         raise NotImplementedError
 
-    def _pkey_auth(self, password=None):
+    def _pkey_auth(self, pkey_file, password=None):
+        raise NotImplementedError
+
+    def _open_session(self):
         raise NotImplementedError
 
     def open_session(self):
@@ -500,9 +503,7 @@ class BaseSSHClient(object):
         raise NotImplementedError
 
     def _sftp_put(self, remote_fh, local_file):
-        with open(local_file, 'rb') as local_fh:
-            for data in local_fh:
-                self._eagain(remote_fh.write, data)
+        raise NotImplementedError
 
     def sftp_put(self, sftp, local_file, remote_file):
         raise NotImplementedError