+++ Changelog.rst
@@ -1,6 +1,24 @@
 Change Log
 ============
 
+2.5.0
++++++
+
+Changes
+-------
+
+* Python 2 no longer supported.
+* Updated class arguments, refactor for ``pssh.clients.native.tunnel``.
+
+Fixes
+-----
+
+* Closed clients with proxy host enabled would not shutdown their proxy servers.
+* Clients with proxy host enabled would not disconnect the proxy client on ``.disconnect`` being called.
+* Default identity files would not be used when private key was not specified - #222.
+* ``ParallelSSHClient(<..>, identity_auth=False`` would not be honoured.
+
+
 2.4.0
 +++++